Add Methodology Page as it's own route on WOW (#224)
As per @toolness's and my observations, it's better practice for us to keep content for the WOW methodology page on Contentful to allow localization, as opposed to a Medium post which doesn't allow that feature. Therefore, we need to bring the methodology page back as a WOW route, and host the text in Contentful.