Allow for age to be null. (#334)
While WoW never seems to ping the `/api/aggregate` endpoint for BBLs that have no HPD registration, DAP does (albeit at the synonymous `/api/dap-aggregate` endpoint), and it sometimes asks about buildings that apparently either have no PLUTO data, or for which PLUTO reports the building's age as being `NULL`.  Regardless, the value our NYCDB returns is `NULL`, which our typecasting code introduced in #326 barfs on, so this fixes it.