Improve error handling. (#380)
This adds a new module called `error-reporting.ts` which introduces a new function, `reportError()`, that we should use to report errors from now on (rather than calling `window.Rollbar.error()` or `console.error()` directly).  In doing so, it decouples us a bit more from Rollbar.

It also fixes #378, in part by adding a new `NetworkError` exception (it looks like the default errors raised by `fetch` are just `TypeError` which is awfully ambiguous).  A `NetworkError` has a `shouldReport` property which defaults to false, but it can be changed to true if it's something that we'd like to know about.  For example, we don't want to report HTTP 500 errors from our own API servers because we know they will report the errors to us from their end, and with more helpful contextual information.  But we _do_ want to know about HTTP 400 errors, because in this case we're somehow issuing bad requests to the API server.