React-only modify-state code and tooltip triangle styling
I changed the code to only render object based on the react component's "this.state". I also modified some styling for the tooltip triangle that goes below the info button. The triangle no longer disturbs the position of other elements in the list, such as "how to use" and "about"