Return HTTP 404 for nonexistent files under /static/. (#401)
It seems #169 re-emerged when we moved the front-end over to Netlify, because Netlify is currently returning the WoW index page with a status of 200 for _every_ URL, including static assets.  This modifies our `_redirects` to actually return 404 for such files, which should hopefully result in less spurious syntax errors coming from Rollbar.