Clean up DetailView component  (#408)
This PR fixes #409 and #410— namely, it refactors the DetailView component like we did other AddressPage child components so we don't have to pass it the addr and addrs props down from its parent. It also introduces a new handler called onOpenDetail that makes sure to bring in the DetailView on mobile when a user mounts the address page for the first time.