Assign and fix type of `allcontacts` field in front-end and back-end (#478)
This PR adds type casting in our back-end python code and and new type definitions to the front-end APIDataTypes.txt file to accommodate the new allcontacts feild added to the wow_bldgs table. In doing so, this PR also adds some tooling to format this new column in the "Export Data" feature.