Add hamburger menu to header on mobile (#493)
This PR replaces our standard navbar of links in the header with a hamburger menu once the viewport is smaller than a desktop or landscape tablet. The hamburger menu includes all original links from the navbar, with the only difference being that instead of the EN/ES toggle, we now show the full language name of whatever language the user is not currently seeing on the site (see the new <LocaleSwitcherWithFullLanguageName> component). I also added an overlay to gray out the main content of the page when the user has the hamburger menu open.