Remove widows from title link on navbar (#497)
This PR makes use of the widont package (that we use on the justfix-website) to remove any single-word second lines from the title link on the header nav bar. This problem specifically arose with the Spanish title as it was too long to fit on one line. Now, instead of looking awkward with a trailing "York" on the second line, it looks much better.