Add Feature Callout widget to the app header menu (#393)
This PR builds on the work of @Sjones21 in that it implements a "Feature Callout" widget that users can access via an icon from the header nav bar. When clicked, a dropdown menu will appear that walks users through a series of GIFs and descriptions that highlight new changes to the tool. 

In addition to Sydney's work, @sraby added a feature flag ENV variable that allows us to enable/disable the presence of this widget without writing new code.