Add search address and detail address nodes to portfolio graph visualization (#547)
This PR refactors our <PortfolioGraph> component to dynamically add on nodes to the graph visualization for the user's Search Address, as well as the Detail Address if it differs from the Search Address. The idea behind these additional nodes was to create some feature parity with the "How is this building associated..." feature on the old WOW:

Now, when users click on this link, they are redirected to the Portfolio Graph (currently on the Summary tab) to see how their search address is related to the address in focus on the Overview tab:

This is particularly important because in order to build trust with our users regarding this new portfolio method, we need to have some feature that illustrates how each building in the portfolio connects back to the original address they searched.