Fix tabbing issues with homepage search bars (#574)
This PR fixes two issues that occur when trying to use the tab key to navigate around the search bars on the homepage:

Once focusing on the Address searchbar, users would be trapped in focus on the search bar even with no results showing up. Now, we only enable a focus trap when there is at least 1 search result in the list to tab through

When closing the Landlord searchbar, focus now returns to the element it previously was on and doesn't get completely lost as it did before.