add Algolia objectID as hash of contents + tests (#588)
this PR adds the unique objectID for each record that Aloglia requires. It is created as a hash of the rest of the contents of the record ({"portfolio_bbl": 1234567890, "landlord_names": "name1, name2"}), so that later we can use the hash to determine if we need to update the record, and hopefully save us some Algolia operations/credits.

Also add tests of this new hash, and adjust the existing tests now that we can ensure consistent ordering of landlord names in the concatenated string.