Ensure that only one landlord corresponds to a given `registrationid` from HPD (#607)
This PR creates an interim fix for the issue of bbls belonging to multiple landlord portfolios on WOWZA. So, we changed our SQL query in generating the raw data for the networkx graph to make sure that only one landlord contact corresponds to each registrationid from HPD. In the rare cases where there are multiple owners on a given registration, we have a ranking for prioritizing which owner type we define as the best one to use for portfolio generation.

Given that the vast majority of these edge casey buildings belong to a portfolio of 1, these changes likely will not affect the scale of our portfolio generation very much, even though we are trimming down the level of possible links a building can have with others. See the comments section of this Shortcut card for more analysis.