Make sure network viz redistributes nodes after adding detail address (#622)
When a user selects a new address from the overview tab or portfolio tab, that triggers the addition of a new node to our network viz to show which landlord name owns that new selected property. Before, this new node would by default render in the center of the graph, with a long overlapping edge connecting it with wherever it's related landlord name node is. This would create a very awkward and busy looking graph, so this PR makes sure that we re-render the graph layout whenever the user's selected detail address changes.

This PR also reduces edge overlap overall and makes sure our node labels are visible from more zoom extents.