link from number of buildings to portfolio tab (#631)
On an address page there is a line at the top saying "your search address is associated with X other buildings" with the number underlined, and it feels like you should be able to click it but it wasn't a link. This PR makes it a link to open the Portfolio tab, and adds a separate google tag for the link.