Add OCA eviction filings to frontend (#655)
This PR finishes up the project of adding OCA (office of court administration) eviction filings data to WOW. First the backend work was done in #642 to extract the necessary data from the private DB maintained by BetaNYC for the address-level version of the housing court data . Then those changes were incorporated into the auto-updating NYCDB (see JustFixNYC/nycdb-k8s-loader#115 & JustFixNYC/nycdb-k8s-loader#117). This PR deals with getting the data from those new database tables and updating the UI to incorporate the new data. It follows the same process as for executed evictions on the overview page, and other indicators on the timeline page, with the one difference being that we can't report filings data on buildings with fewer than 11 units so there is a difference between 0 filings and N/A and this is noted in the timeline graph.