upgrade python to 3.9.16 (#694)
We need to upgrade our Heroku stack and the new stack doesn't support python 3.6 anymore, so this PR upgrades python to version 3.9.16. Heroku recommended 3.10.9, but I kept running into problems with the invalid wheels from numpy and found things online saying they haven't made 3.10 wheels available yet, so I settled on the other recommended version - 3.9.16. Then there were a few package upgrades that were required to work with the newer python.

The new stack also no longer includes python2, and node-sass requires that, but the package is also deprecated now, so we are also updating to the new sass module.