upgrade typescript to ^4.9.3 (#697)
We need to upgrade typescript to be able to add some of the new dependencies for portfolio filters, like tanstack/react-table, and also for the new amplitude sdk. This PR upgrades typescript to ^4.9.3, and then makes a variety of small changes and other package upgrades to address all the errors that came up. This includes upgrading react-scripts and changing the import of inline svg.