portfolio table filters (#700)
Initial implementation of the portfolio filters for table on desktop. This sets up the react Context to share filter options and selections between the filters bar component and the table (and later the map). It upgrades from the older react-table v6 to TanStack Table v8. Adds a multiselect component adapted from multiselect-react-dropdown, as well as a binary toggle and min/max numeric input. It also addres a details/summary dropdown component to contain the filter components. Along with the new v8 table, this adds the filter logic to update the columns. The filters we have are for presence of any rent stabilized (toggle), landlord name (multiselect), units in the building (min/max), and zipcode (multiple select). The "landlord" column/filter now includes all the different types of contacts (owner, corporation, head officer, etc.) in both the filter, and in a new expanding row for the column/table.