Filters mobile (#718)
This PR adds support for mobile to the filter bar. Makes the basic changes to the styles of filter components and the bar, plus adds the fullscreen wrapper of all filters - and the scroll snapping and shadow effects. Plus a small fix to the handling of min-max errors, and its default values I caught while in here.

Also included is an update to the filter input styling on desktop and mobile so they read more clearly as inputs (not full width, white background)