Font weight and spacing cross browser consistency (#740)
There has been a very noticeable difference in font-weight and letter-spacing across browsers.

Chrome: already pretty much as intended.
Safari and Firefox: both the font-weight appeared much higher, and the letter spacing was larger.

I made sure that we have the specific weights specified when loading the font files. I also added -webkit-font-smoothing: antialiased, and this solves the weight issue in Safari, but if applied in Chrome it makes it too bold, and does nothing for Firefox. This also didn't help the letter spacing.

I kept trying to solve this with these sorts of browser-independent fixes, but have just not been able to find something that solves all of them.

So I've tried to instead use JS to figure out the browser for the user-agent on start up and store that in a data attribute in the html. Then in SASS we can use a mixin to give browser-specific styles, like we do for screen sizes.

I know it's not best practice to try to infer browser from user agent to change styles, and but since all the global fixes I could find didn't work, and it's not a situation where we could specifically test for availability of some css property, I can't think of how else to deal with his. Also, since the changes are not critical, the risk seems low.

I use these browser-specific mixins to adjust the font weight and letter spacing for the remaining issues.