Minmax select refactor and add presets (#749)
A big set of changes to our building size filter.

The data structure of the options/selections ranges is changed from [number, number] to a more sensible [{min: number, max: number}], (an array because we can no have multiple ranges - see below). Also instead of using undefined as the default value, it uses -Infinity/Infinity (the undefined were changed to this anyway in the table filter application)
The error handling for the min/max inputs also uses a similar data structure (plus a msg field) for better readability
There are now a set of preset unit ranges to choose from (can be multiple), or you can expand a dropdown to reveal the original min/max inputs to define a custom range. When you enter anything in the custom inputs it disables the presets (but remembers your selections if you then clear the inputs). There's still the one apply button for both options.
There are also a lot of a11y fixes in here: input options following these suggestions, proper labels and descriptions for screen readers (including the new presets), focus states, alert message.