owner filter info modal pauses accordion focus trap (#746)
This PR fixes an issue with the landlord filter accordion that when the info modal open focus stays trapped in the accordion. In a newer version of react-focus-trap (I had to update the package) there is a pause prop to deal with this sort of situation. However this required that I have access to the open/closed state of the modal, so instead of passing the callback from the parent, we pass the contents of the modal and create it within the accordion component. There are also some copy changes. Also, seeing that the text for the info button is now custom for the filter topic (not the generic "what's this") I added that as a prop for the filter accordion.