fix hidden underline on IOS (#756)
For the design system we use text-underline-position: under for everything so that the underline never crosses descenders in the text. This has caused some issues (on IOS only it seems) because the auto height of the element content doesn't account this lower underline, it just uses the line-height as normal, and there are inconsistencies with whether the underline shows above other content. On IOS if display is block or inline-block it doesn't show through, but it does if it's flex. And it just seems to work regardless on other browser/devices.