add analytics events for filters (#752)
This PR adds a collection of analytics events - some are already in wow but needed to be added back after the table was completely rewritten for filters, and others are brand new to cover the filters interactions. All of the events have parity between Google and Amplitude, including the custom properties we send along with each (eg. portfolioSize, portfolioColumn, etc.). [sc-10784]

The way I'm actually logging the events in kind of awkward, because we have some common extra properties to send with them all, I tried to make them not too clunky when actually doing the logging, and we also have some frustrating inconsistencies between amplitude and google. Once we have everything up for filters, and can spend a bit of time getting settled into the new Google ecosystem of analytics tools, we'll have a better idea of what to change (eg. maybe we can drop amplitude all together?). These things should be addressed in a update of all the analytics event setup for WOW - I've started adding some thoughts on that in shortcut [sc-12296]

There is also a small bug that I noticed while testing - when there are 0 records for the table (from filtering) it showed up as 0.