I18n fixes (#753)
Previously (#743) I tried to remove i18n as a prop from some of the filters components and instead just import it in those files. However, I misunderstood exactly how this is supposed to work and since we didn't have all the translations in I didn't notice that it broke at first.

This PR fixes those issues using the proper ways to pass around the correct i18n prop without having to manually add it each time, following the examples elsewhere in wow. This makes use of <I18n> {({i18n}) => ()} </I18n> pattern for the multiselect and minmaxselect components, and the withI18nProps & withI18n() pattern for the Filters and Table components.