remove auto focus of inputs on opening accordions (#759)
We've gone back and forth on this a few times. Most recently, on desktop only, when you opened the filter accordions it would automatically focus the inputs. This had already been disabled for mobile, and now this behavior has been removed from desktop as well. (Though the optional prop is retained in FilterAccordion in case we change our minds again later)