Clear unapplied filter selections (#764)
For our multiselect filters all selections made from the options dropdown are saved in the multiselect component state as "pending" selections, then once "apply" is clicked that are confirmed and passed into the main filters context/state. We want to make sure things match user expectations, so this PR changes so that:

any selections that are not applied are cleared when the filter dropdown is closed
any applied selections persist in the filter when opened/closed
when all applied filter selections are cleared with the "clear all" button the "pending" selections should also be cleared from the multiselect component.

Also, in doing this I realized there was some inconsistency with the types of the multiselect filter selections that needed to be fixed. No more any used in those!

For the MinMax select filter we do the same thing. However, to be able to show the applied selections in the component based on the applied filter selections (since they are cleared from the internal minmax select state on every close) we needed to change the data structure of those applied selections to know whether the selection is from the custom range or the presets (or if it just the default -Infinity/Infinity min max values).