Focus inputs and keyboard display shifts entire HTML tag on iOS only (#769)
any time the keyboard view is shown, the HTML background (table hidden behind mobile filter view) would shift in weird ways. the fix is calling a force scroll function before the user navigates away from the full-screen filter view. currently i call the function on "view results" button and "close" button, since those are the only pathways out of the full-screen filter. however if we were to generalize the Multiselect and MinMaxSelect components, we should call this same function whenever the input focus is triggered to avoid the buggy behavior!