fix broken column sorting in portfolio table (#771)
When adding a analytics event on the column sort click in #752 I accidentally broke the table sorting. There is a method from the react-table that handles the sorting - but it actually returns a handler function rather than actually handling it. So I just needed to actually call it after getting the handler.