Add clear selections button to min-max filter component (#772)
This adds a "clear selections" button to the min-max select filter component for buildings size. It only appears when there are active selections (including if they are still just "pending", unapplied selections). Clicking it clears both the preset selections and the custom range and closes the filter. This is the same behavior as the clear button on the multiselect - it works like the "apply" button. We also a chevron icon for the custom range dropdown. The html needed a slight restructuring to get the alignment right, so there is some extra nesting of the styles, but there aren't major changes there. The width of the desktop filter needs to be expanded to fit the extra text.