ios mobile menu scrolling (#774)
We have been having problems with the mobile filter menu on IOS cause by layout shifts when the virtual keyboard appears. The way IOS (Chrome and Safari) handles the changing space is to keep the page the same size, but push it up and off screen at the top.

This is their intended behavior, but it has been causing some bugs as well. First, it would cause a permeant shift of the page that would persist even after closing the keyboard and exiting the mobile filter menu - leaving the whole portfolio table page with a black/white section at the bottom of the page. We found a workaround for this in #769, using JS to scroll the page once the filter menu is closed, and this fixed the layout back on the portfolio table page.

Another problem we were having is that on the menu when the keyboard is active, you couldn't scroll all the way down - it would just get stuck. This meant that you couldn't make some selections from the multiselect dropdown then scroll down to click apply, without first exiting the keyboard.

This PR makes some CSS changes for the mobile filter menu that solves both of these problems (so we can remove the JS scroll functions), for IOS Safari and Chrome (and it all still works properly, the same as before, for the other browsers/devices). We change the heights of the <details/> and the <div/> within it so that they use 100% instead of 100vh, use absolute instead of fixed positioning for details, and use default positioning for the inside div. Finally, for all this to work properly we need to make the parent of the absolute positioned details, which is the body, positioned relative.

This leaves just one remaining issue for IOS Safari. Because when the keyboard it up and the page is intentionally pushed off screen, a new level of scrolling is added so you can scroll the full page to see the areas pushed off screen. Since we already have scrolling on the filter menu, once this extra level is added it means that you have the keyboard up and first scroll the menu normally to the bottom, then if you try to scroll down more it actives the extra page scrolling and you can pull up this weird empty black bar. If you close the keyboard, every thing goes back to normal for scrolling and the black bar disappears.

I've found a ton of different questions about this online (going back years) but no solutions.

Since this is still an improvement over what we had, I'm going to merge this in so we can move ahead with other design review, and will continue looking into this black bar issue separately.

Also in this PR, I change the z-indexes on the mobile menu for fore sensible values, remove a callbacks for the min-max select on focus of the inputs since now the OS handles the scroll to keep them in view as the keyboard appears.