add resize-observer-polyfill package (#801)
In #787, when we had problems with the portfolio tab map not adjusting when the container height changed, we introduced a new package to use ResizeObserver. It is apparently supported by ~95% of browsers, but in the few cases it's not supported we are getting errors that it's not defined.

A github issues thread on the package repo explains that there is an alternative (the library resize-observer-polyfill) that can be assigned to window.ResizeObserver to handle these cases.

I tested this out by just first setting window.ResizeObserver = undefined, confirming it breaks, then using the polyfill and confirming it works again, and haven't had any problems.