remove alert on overview map to filter portfolio (#844)
When we launched portfolio filters we added an alert on overview map that prompts users to check out the portfolio tab to filter down the properties. We are now seeing extremely low rates of clicks on this alert, so are going to remove it along with the launch of email alerts.