Add validation/error handling to account settings update credentials (#854)
This PR adds input validation and error handling to the forms for updating user email and password on the account settings page to ensure the new email is valid and not already in use, and that the new password satisfies rules. It swaps in the new <EmailInput>, and adapts the existing validation/error handling logic from ` to show input- and page-level alerts. It prevents the default form reload and controls the toggling of edit mode after successful submission. Also adds an option to show an input-level error for password when rules aren't shown. Also fixes some spacing/styles that might have gotten messed up with the modal login changes. Also adds an input-level error state for password when the rule validation messages aren't shown (eg. for empty input when logging in).
There is now a "callout" box under email when not editing if you haven't verified your email. And when you successfully change your email we mark it un-verified and send a verify email on the server side (JustFixNYC/auth-provider#17) and then the callout appears.