add component library buttons, fix verified condition on building page (#873)
I was starting to implement the component-library buttons, then Leslie flagged that the condition was wrong on the building pages. When you're not verified you should always see the prompt to verify and not allow users to subscribe until they are. previously it was allowing them to subscribe then the prompt to verify would show up after. So this fixes that plus a few button swaps that will get continued in another PR.