don't clear errors on login/signup toggle, only hide them (#876)
When toggling between login/signup we had issue of errors persisting, then in #865 we cleared them on toggle. But this created a new bug that allowed you to create an account with a password not meeting rules if you first toggled between login/signup. Now we just turn off the showing of errors on toggle, but don't clear the errors themselves so submit is still prevented.