update HPD Online link for bbl search results, useful links and alerts api (#889)
When the new HPD Online first launched it wouldn't allow you to link to search results based on BBL like the old version. This is important because since WOW uses BBLs for our pages, HPD uses their own "HPD building ID" and a single BBL can have multiple buildings. When there is only one building we can know that and link directly to the page, but when there are multiple we need to link to these bbl search results and the user can select the building they are interested in.
Now that HPD allows these links, we update our own "useful links" section.
Also in this PR I add this link (single building or search results) as an optional "indicator" on the email alerts endpoint so that we can include those in the building update emails just like useful links.
The corresponding PR that makes use of the new api changes: JustFixNYC/auth-provider#34