add verify step to nav signup flow, same as in modal (#891)
Previous the signup flow from Nav would end after user type selection and redirect users to the account settings page. This was confusing during initial user testing since it never said anything about email verification until later it was show on building pages, also there is no need to do anything on account settings after sign up. This PR removes that redirect and shows the same "verify email" prompt as we show at the end of the on-page/in-modal sign up flow. So the final step looks like this.