Standalone page (#893)
This adds a standardized "standalone" page design for login/signup, forgot-password, reset-password, verify email (plus a modified version for unsubscribe/manage subscriptions).
[sc-14218]

This new page also gets a new minimal footer.
[sc-13964]

Mobile styles are also fixed for all these pages with the new layout.
[sc-14066]

All automatic/countdown redirects are removed from login/signup via nav, reset-password and verify.
[sc-14144]

The login flow from nav now has an extra final step. If not verified you get the same prompt to do so as at the end of sign up, and it sends a new email. And if you are already verified then you get a a new "success" step with a link to the homepage.

Focus now automatically shifts to the password input after submitting email in the login/signup flows.
[sc-14271]

Also, I realized there was a bad copy/paste error on the reset-password page - it was sending a verify email instead of a reset one! So I fixed that and also realized there was a problem with the request because cookies weren't getting sent and those are needed for accessing session on the auth server.

I also changed the JF logo svg (also used in nav) to a component so that we can make use of the "currentColor" to style with css instead of adding a second black version.