Simplified login/signup flow, removes modal on building page (#904)
This PR removes the entire on-building-page login/signup flow and modal. Now there is a single standalone login/signup page and when you're on a building page and click the "add building" button if you aren't logged in it just sends you to that page. We now make use of react-router's location state to set the building page's AddressRecord as state so we can have access to that from within the standalone login/signup flow. This allows us to still subscribe onSuccess of the login/register call, include the address subHeaders, provide a link back to the address page from the ending VerifyEmail step of those flows, or to automatically redirect users back to the building page on successful login for already verified users. We also make use of the same location state approach to set justSubscribed when landing back on the address page so that we can ensure it shows subscribed state immediately and we can know to show the page-level alert notifying the user that they are logged in and subscribed. Because the location state persists after reloads, we transfer the data to regular component state within a useEffect then clear the location state on both login and building page.

Also adds a loading state to all submit buttons in login/signup flow
Also ensures that emails are not broken by wrapping in login/signup flow

This also fixes the bug where after login/sign to subscribe to a building, the user context was out of sync and still showed the user as not logged in on the nav etc. It seems that this was due userContext.subscribe() overwriting the user values (email, verified) with the default undefined state because of the same kind of bug we had previous where an empty JustfixUser object is truthy, so you need to use !!user?.email in conditions instead (see shortcut story for details).