add api endpoint for dataset last updated (#949)
Pairs with JustFixNYC/nycdb-k8s-loader#159 which adds a table with every dataset and when it was last updated in our db. This adds a simple endpoint that takes a dataset name and returns the date it was last updated. We'll use this to provide indicator-specific last updated dates in the Signature dashboard, and it should also help for general monitoring now that there are so many dataset in there and so many are now dependencies for wow/alerts/signature.

I haven't added tests yet because I would need to add a test data builder and ran out of time. I also need to add signature api tests, so I'll do all of this as a post mvp launch task.