Signature Dashboard backend (#946)
This adds everything for the backend of the Signature Dashboard.

Follows the example of ocaevictions, adding a new signature folder to install on the k8s-loader for a signature job on kubernetes. This fetches a csv from our AWS S3 that contains all the data provided by UNHP and uploads that to the DB. Then there are a number of SQL scripts that generate the tables that power the api. Specifically, signature_buildings, signature_building_charts, signature_collections, and signature_collection_charts.

Then it also adds new API endpoints to be used in the dashboard. All of the important logic for creating the indicators is in the script creating the tables, so the api queries are very simple.